feat: add initial (and experimental) support for ZFS
On config creation, a new dataset is created (`.snapshots`) that will
store snapshots information and a symlink to the snapshot in the `.zfs`
directory.  This dataset is destroyed on config deletion.

Due to a ZFS “limitations”, snapshots can only be created read-only and
cannot be derived from a parent.

This is based on the initial idea and work of Daniel Sullivan
(mumblepins).

Implements #145.